Split tests and benches to compile on stable
