+++ Cargo.toml
@@ -5,15 +5,22 @@ edition = "2018"
 
 [lib]
 name = "sc2pathlib"
-crate-type = ["cdylib"]
+crate-type = ["cdylib", "rlib"]
 
 [dependencies.pyo3]
-version = "0.10.1"
+version = "0.11.0"
 features = ["extension-module"]
 
 [dependencies]
 pathfinding = "2.0.4"
 
+[dev-dependencies]
+criterion = "0.3"
+
+
+[[bench]]
+name = "path_benchmark"
+harness = false
 # [profile.release]
 # opt-level = 3
 # debug=false
